requirements.txt
appdirs==1.4.4
certifi==2025.1.31
charset-normalizer==3.4.1
docker==7.1.0
idna==3.10
lxml==5.4.0
packaging==25.0
psutil==7.0.0
pyshark==0.6
python-dotenv==1.1.0
requests==2.32.3
termcolor==3.0.1
urllib3==2.4.0
